Enum tidb_query_datatype::codec::batch::LazyBatchColumn

pub enum LazyBatchColumn {
    Raw(BufferVec),
    Decoded(VectorValue),
}

A container stores an array of datums, which can be either raw (not decoded), or decoded into the VectorValue type.

TODO: Since currently the data format in response can be the same as in storage, we use this structure to avoid unnecessary repeated serialization / deserialization. In future, Coprocessor will respond all data in Chunk format which is different to the format in storage. At that time, this structure is no longer useful and should be removed.

Variants

Raw(BufferVec)

Decoded(VectorValue)

Implementations

impl LazyBatchColumn

pub fn raw_with_capacity(capacity: usize) -> Self

Creates a new LazyBatchColumn::Raw with specified capacity.

pub fn decoded_with_capacity_and_tp(capacity: usize, eval_tp: EvalType) -> Self

Creates a new LazyBatchColumn::Decoded with specified capacity and eval type.

pub fn clone_empty(&self, capacity: usize) -> Self

Creates a new empty LazyBatchColumn with the same schema.

pub fn is_raw(&self) -> bool

pub fn is_decoded(&self) -> bool

pub fn decoded(&self) -> &VectorValue

pub fn mut_decoded(&mut self) -> &mut VectorValue

pub fn raw(&self) -> &BufferVec

pub fn mut_raw(&mut self) -> &mut BufferVec

pub fn len(&self) -> usize

pub fn truncate(&mut self, len: usize)

pub fn clear(&mut self)

pub fn is_empty(&self) -> bool

pub fn capacity(&self) -> usize

pub fn ensure_decoded(
&mut self,
ctx: &mut EvalContext,
field_type: &FieldType,
logical_rows: &[usize]
) -> Result<()>

Decodes this column if the column is not decoded, according to the given logical rows map. After decoding, the decoded column will have the same physical layout as the encoded one (i.e. the same logical rows), but elements in unnecessary positions will not be decoded and will be None.

The field type is needed because we use the same DateTime structure when handling Date, Time or Timestamp.

pub fn ensure_all_decoded_for_test(
&mut self,
ctx: &mut EvalContext,
field_type: &FieldType
) -> Result<()>

pub fn maximum_encoded_size(&self, logical_rows: &[usize]) -> usize

Returns maximum encoded size.

pub fn maximum_encoded_size_chunk(&self, logical_rows: &[usize]) -> usize

Returns maximum encoded size in chunk format.

pub fn encode(
&self,
row_index: usize,
field_type: &impl FieldTypeAccessor,
ctx: &mut EvalContext,
output: &mut Vec<u8>
) -> Result<()>

Encodes into binary format.

pub fn encode_chunk(
&self,
ctx: &mut EvalContext,
logical_rows: &[usize],
field_type: &FieldType,
output: &mut Vec<u8>
) -> Result<()>

Encodes into Chunk format.