Struct tidb_query_datatype::codec::batch::LazyBatchColumnVec

pub struct LazyBatchColumnVec {
    columns: Vec<LazyBatchColumn>,
}

Stores multiple LazyBatchColumns. Each column has an equal length.

Fields
columns: Vec<LazyBatchColumn>

Multiple lazy batch columns. Each column is either decoded, or not decoded.

For decoded columns, they may be in different types. If the column is in type LazyBatchColumn::Raw, it means that it is not decoded.

Implementations

impl LazyBatchColumnVec

pub fn empty() -> Self

Creates a new empty LazyBatchColumnVec, which does not have columns and rows.

Because column numbers won’t change, it means constructed instance will be always empty.

pub fn clone_empty(&self, capacity: usize) -> Self

Creates a new empty LazyBatchColumnVec with the same number of columns and schema.

pub fn columns_len(&self) -> usize

Returns the number of columns.

It might be possible that there is no row but multiple columns.

pub fn rows_len(&self) -> usize

Returns the number of rows.

pub fn assert_columns_equal_length(&self)

Asserts that all columns have equal length.

pub fn maximum_encoded_size(
&self,
logical_rows: &[usize],
output_offsets: &[u32]
) -> usize

Returns maximum encoded size.

pub fn maximum_encoded_size_chunk(
&self,
logical_rows: &[usize],
output_offsets: &[u32]
) -> usize

Returns maximum encoded size in chunk format.

pub fn encode(
&self,
logical_rows: &[usize],
output_offsets: &[u32],
schema: &[FieldType],
output: &mut Vec<u8>,
ctx: &mut EvalContext
) -> Result<()>

Encodes into binary format.

pub fn encode_chunk(
&mut self,
logical_rows: &[usize],
output_offsets: &[u32],
schema: &[FieldType],
output: &mut Vec<u8>,
ctx: &mut EvalContext
) -> Result<()>

Encode into chunk format.

pub fn truncate_into_equal_length(&mut self)

Truncates columns into equal length. The new length of all columns would be the length of the shortest column before calling this function.

pub fn as_slice(&self) -> &[LazyBatchColumn]

Returns the inner columns as a slice.

pub fn as_mut_slice(&mut self) -> &mut [LazyBatchColumn]

Returns the inner columns as a mutable slice.