Module tidb_query_datatype::codec::chunk

Re-exports

pub use crate::codec::Error;

pub use crate::codec::Result;

Modules

chunk

column

Structs

Chunk
Chunk stores multiple rows of data. Values are appended in compact format and can be directly accessed without decoding. When the chunk is done processing, we can reuse the allocated memory by resetting it.

Column
Column stores the same column data of multi rows in one chunk.

Traits

ChunkColumnEncoder
ChunkColumnEncoder encodes the Chunk column.

ChunkEncoder
ChunkEncoder encodes the chunk.