Function tidb_query_datatype::codec::convert::float_str_to_int_string

fn float_str_to_int_string<'a>(
    ctx: &mut EvalContext, 
    valid_float: &'a str
) -> Result<Cow<'a, str>>

It converts a valid float string into valid integer string which can be parsed by i64::from_str, we can't parse float first then convert it to string because precision will be lost.

When the float string indicating a value that is overflowing the i64, the original float string is returned and an overflow warning is attached.

This func will find serious overflow such as the len of result > 20 (without prefix +/-) however, it will not check whether the result overflow BIGINT.