Module tidb_query_datatype::codec::data_type

Re-exports

pub use crate::codec::mysql::json::JsonRef;

pub use crate::codec::mysql::Decimal;

pub use crate::codec::mysql::Duration;

pub use crate::codec::mysql::Json;

pub use crate::codec::mysql::JsonType;

pub use crate::codec::mysql::Time as DateTime;

Modules

bit_vec

chunked_vec_bytes

chunked_vec_common

chunked_vec_json

chunked_vec_sized

logical_rows

scalar

vector

Structs

BitAndIterator

BitVec
A boolean vector, which consolidates 64 booleans into 1 u64 to save space.

BytesGuard

BytesWriter

ChunkedVecBytes

ChunkedVecJson
A vector storing Option<Json> with a compact layout.

ChunkedVecSized
A vector storing Option<T> with a compact layout.

PartialBytesWriter

Enums

LogicalRows
LogicalRows is a replacement for logical_rows parameter in many of the copr functions. By distinguishing identical and non-identical mapping with a enum, we can directly tell if a logical_rows contains all items in a vector, and we may optimiaze many cases by using direct copy and construction.

ScalarValue
A scalar value container, a.k.a. datum, for all concrete eval types.

ScalarValueRef
A scalar value reference container. Can be created from ScalarValue or VectorValue.

VectorValue
A vector value container, a.k.a. column, for all concrete eval types.

Constants

BATCH_MAX_SIZE

Statics

IDENTICAL_LOGICAL_ROWS
Identical logical row is a special case in expression evaluation that the rows in physical_value are continuous and in order.

Traits

AsMySQLBool
A trait of evaluating current concrete eval type into a MySQL logic value, represented by Rust's bool type.

ChunkRef

ChunkedVec

Evaluable
A trait of all types that can be used during evaluation (eval type).

EvaluableRef

EvaluableRet

UnsafeRefInto

VectorValueExt
Additional VectorValue methods available via generics. These methods support different concrete types but have same names and should be specified via the generic parameter type.

Type Definitions

Bytes

BytesRef

Int

Real