Module tidb_query_datatype::codec::data_type::logical_rows

Structs

LogicalRowsIterator

Enums

LogicalRows
LogicalRows is a replacement for logical_rows parameter in many of the copr functions. By distinguishing identical and non-identical mapping with a enum, we can directly tell if a logical_rows contains all items in a vector, and we may optimiaze many cases by using direct copy and construction.

Constants

BATCH_MAX_SIZE

Statics

IDENTICAL_LOGICAL_ROWS
Identical logical row is a special case in expression evaluation that the rows in physical_value are continuous and in order.