Struct tidb_query_datatype::codec::data_type::ChunkedVecSized

pub struct ChunkedVecSized<T: Sized> {
    data: Vec<T>,
    bitmap: BitVec,
    phantom: PhantomData<T>,
}

A vector storing Option<T> with a compact layout.

T must be a primitive structure. All data must be stored in that structure itself. This includes Int, Real, Decimal, DateTime and Duration in copr framework.

Inside ChunkedVecSized, bitmap indicates if an element at given index is null, and data stores actual data. If the element at given index is null (or None), the corresponding bitmap bit is false, and data stores zero value for that element. Otherwise, data stores actual data, and bitmap bit is true.

Fields
data: Vec<T>bitmap: BitVecphantom: PhantomData<T>
Implementations

impl<T: Sized + Clone> ChunkedVecSized<T>

fn get(&self, idx: usize) -> Option<&T>