Enum tidb_query_datatype::codec::data_type::vector::VectorValue

pub enum VectorValue {
    Int(ChunkedVecSized<Int>),
    Real(ChunkedVecSized<Real>),
    Decimal(ChunkedVecSized<Decimal>),
    Bytes(ChunkedVecBytes),
    DateTime(ChunkedVecSized<DateTime>),
    Duration(ChunkedVecSized<Duration>),
    Json(ChunkedVecJson),
    Enum(ChunkedVecEnum),
    Set(ChunkedVecSet),
}

A vector value container, a.k.a. column, for all concrete eval types.

The inner concrete value is immutable. However it is allowed to push and remove values from this vector container.

Variants

Int(ChunkedVecSized<Int>)

Real(ChunkedVecSized<Real>)

Decimal(ChunkedVecSized<Decimal>)

Bytes(ChunkedVecBytes)

DateTime(ChunkedVecSized<DateTime>)

Duration(ChunkedVecSized<Duration>)

Json(ChunkedVecJson)

Enum(ChunkedVecEnum)

Set(ChunkedVecSet)

Implementations

impl VectorValue

pub fn with_capacity(capacity: usize, eval_tp: EvalType) -> Self

Creates an empty VectorValue according to eval_tp and reserves capacity according to capacity.

pub fn clone_empty(&self, capacity: usize) -> Self

Creates a new empty VectorValue with the same eval type.

pub fn eval_type(&self) -> EvalType

Returns the EvalType used to construct current column.

pub fn len(&self) -> usize

Returns the number of datums contained in this column.

pub fn is_empty(&self) -> bool

Returns whether this column is empty.

Equals to len() == 0.

pub fn truncate(&mut self, len: usize)

Shortens the column, keeping the first len datums and dropping the rest.

If len is greater than the column's current length, this has no effect.

pub fn clear(&mut self)

Clears the column, removing all datums.

pub fn capacity(&self) -> usize

Returns the number of elements this column can hold without reallocating.

pub fn append(&mut self, other: &mut VectorValue)

Moves all the elements of other into Self, leaving other empty.

Panics

Panics if other does not have the same EvalType as Self.

pub fn eval_as_mysql_bools(
&self,
ctx: &mut EvalContext,
outputs: &mut [bool]
) -> Result<()>

Evaluates values into MySQL logic values.

The caller must provide an output buffer which is large enough for holding values.

pub fn get_scalar_ref(&self, index: usize) -> ScalarValueRef<'_>

Gets a reference of the element in corresponding index.

Panics

Panics if index is out of range.

pub fn maximum_encoded_size(&self, logical_rows: &[usize]) -> usize

Returns maximum encoded size in binary format.

pub fn maximum_encoded_size_chunk(&self, logical_rows: &[usize]) -> usize

Returns maximum encoded size in chunk format.

pub fn encode(
&self,
row_index: usize,
field_type: &impl FieldTypeAccessor,
ctx: &mut EvalContext,
output: &mut Vec<u8>
) -> Result<()>

Encodes a single element into binary format.

pub fn encode_sort_key(
&self,
row_index: usize,
field_type: &impl FieldTypeAccessor,
ctx: &mut EvalContext,
output: &mut Vec<u8>
) -> Result<()>

impl VectorValue

pub fn to_int_vec(&self) -> Vec<Option<Int>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_real_vec(&self) -> Vec<Option<Real>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_decimal_vec(&self) -> Vec<Option<Decimal>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_bytes_vec(&self) -> Vec<Option<Bytes>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_date_time_vec(&self) -> Vec<Option<DateTime>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_duration_vec(&self) -> Vec<Option<Duration>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_json_vec(&self) -> Vec<Option<Json>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_enum_vec(&self) -> Vec<Option<Enum>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn to_set_vec(&self) -> Vec<Option<Set>>

Extracts a slice of values in specified concrete type from current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_int(&mut self, v: Option<Int>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_real(&mut self, v: Option<Real>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_decimal(&mut self, v: Option<Decimal>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_bytes(&mut self, v: Option<Bytes>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_date_time(&mut self, v: Option<DateTime>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_duration(&mut self, v: Option<Duration>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_json(&mut self, v: Option<Json>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_enum(&mut self, v: Option<Enum>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.

impl VectorValue

pub fn push_set(&mut self, v: Option<Set>)

Pushes a value in specified concrete type into current column.

Panics

Panics if the current column does not match the type.