Trait tidb_query_datatype::codec::data_type::vector::VectorValueExt

pub trait VectorValueExt<T: EvaluableRet> {
    fn push(&mut self, v: Option<T>);
}

Additional VectorValue methods available via generics. These methods support different concrete types but have same names and should be specified via the generic parameter type.

Required methods

fn push(&mut self, v: Option<T>)

The generic version for VectorValue::push_xxx().

Loading content...