Enum tidb_query_datatype::codec::datum::Datum

pub enum Datum {
    Null,
    I64(i64),
    U64(u64),
    F64(f64),
    Dur(Duration),
    Bytes(Vec<u8>),
    Dec(Decimal),
    Time(Time),
    Json(Json),
    Enum(Enum),
    Set(Set),
    Min,
    Max,
}

Datum stores data with different types.

Variants

Null

I64(i64)

U64(u64)

F64(f64)

Dur(Duration)

Bytes(Vec<u8>)

Dec(Decimal)

Time(Time)

Json(Json)

Enum(Enum)

Set(Set)

Min

Max

Implementations

impl Datum

pub fn as_int(&self) -> Result<Option<i64>>

pub fn as_real(&self) -> Result<Option<f64>>

pub fn as_decimal(&self) -> Result<Option<Cow<'_, Decimal>>>

pub fn as_string(&self) -> Result<Option<Cow<'_, [u8]>>>

pub fn as_time(&self) -> Result<Option<Cow<'_, Time>>>

pub fn as_duration(&self) -> Result<Option<Duration>>

pub fn as_json(&self) -> Result<Option<Cow<'_, Json>>>

impl Datum

pub fn cmp(&self, ctx: &mut EvalContext, datum: &Datum) -> Result<Ordering>

cmp compares the datum and returns an Ordering.

fn cmp_i64(&self, ctx: &mut EvalContext, i: i64) -> Result<Ordering>

fn cmp_u64(&self, ctx: &mut EvalContext, u: u64) -> Result<Ordering>

fn cmp_f64(&self, ctx: &mut EvalContext, f: f64) -> Result<Ordering>

fn cmp_bytes(&self, ctx: &mut EvalContext, bs: &[u8]) -> Result<Ordering>

fn cmp_dec(&self, ctx: &mut EvalContext, dec: &Decimal) -> Result<Ordering>

fn cmp_dur(&self, ctx: &mut EvalContext, d: Duration) -> Result<Ordering>

fn cmp_time(&self, ctx: &mut EvalContext, time: Time) -> Result<Ordering>

fn cmp_json(&self, ctx: &mut EvalContext, json: &Json) -> Result<Ordering>

fn cmp_enum(&self, _ctx: &mut EvalContext, v: &Enum) -> Result<Ordering>

fn cmp_set(&self, _ctx: &mut EvalContext, v: &Set) -> Result<Ordering>

pub fn into_bool(self, ctx: &mut EvalContext) -> Result<Option<bool>>

into_bool converts self to a bool. source function name is ToBool.

pub fn to_string(&self) -> Result<String>

to_string returns a string representation of the datum.

pub fn into_string(self) -> Result<String>

into_string convert self into a string. source function name is ToString.

pub fn into_f64(self, ctx: &mut EvalContext) -> Result<f64>

into_f64 converts self into f64. source function name is ToFloat64.

pub fn into_i64(self, ctx: &mut EvalContext) -> Result<i64>

into_i64 converts self into i64. source function name is ToInt64.

pub fn f64(&self) -> f64

Keep compatible with TiDB's GetFloat64 function.

pub fn i64(&self) -> i64

Keep compatible with TiDB's GetInt64 function.

pub fn u64(&self) -> u64

Keep compatible with TiDB's GetUint64 function.

pub fn into_arith(self, ctx: &mut EvalContext) -> Result<Datum>

into_arith converts datum to appropriate datum for arithmetic computing. Keep compatible with TiDB's CoerceArithmetic function.

pub fn into_dec(self) -> Result<Decimal>

Keep compatible with TiDB's ToDecimal function. FIXME: the EvalContext should be passed by caller

pub fn cast_as_json(self) -> Result<Json>

cast_as_json converts Datum::Bytes(bs) into Json::from_str(bs) and Datum::Null would be illegal. It would be used in cast, json_merge,json_extract,json_type mysql> SELECT CAST('null' AS JSON); +----------------------+ | CAST('null' AS JSON) | +----------------------+ | null | +----------------------+

pub fn into_json(self) -> Result<Json>

into_json would convert Datum::Bytes(bs) into Json::from_string(bs) and convert Datum::Null into Json::none(). This func would be used in json_unquote and json_modify

pub fn to_json_path_expr(&self) -> Result<PathExpression>

to_json_path_expr parses Datum::Bytes(b) to a JSON PathExpression.

fn coerce_to_dec(self) -> Result<Datum>

Try its best effort to convert into a decimal datum. source function name is ConvertDatumToDecimal.

fn coerce_to_f64(self, ctx: &mut EvalContext) -> Result<Datum>

Try its best effort to convert into a f64 datum.

pub fn coerce(
ctx: &mut EvalContext,
left: Datum,
right: Datum
) -> Result<(Datum, Datum)>

coerce changes type. If left or right is F64, changes the both to F64. Else if left or right is Decimal, changes the both to Decimal. Keep compatible with TiDB's CoerceDatum function.

pub fn checked_div(self, ctx: &mut EvalContext, d: Datum) -> Result<Datum>

checked_div computes the result of self / d.

pub fn checked_add(self, _: &mut EvalContext, d: Datum) -> Result<Datum>

Keep compatible with TiDB's ComputePlus function.

pub fn checked_minus(self, _: &mut EvalContext, d: Datum) -> Result<Datum>

checked_minus computes the result of self - d.

pub fn checked_mul(self, _: &mut EvalContext, d: Datum) -> Result<Datum>

pub fn checked_rem(self, _: &mut EvalContext, d: Datum) -> Result<Datum>

pub fn checked_int_div(self, _: &mut EvalContext, datum: Datum) -> Result<Datum>