Module tidb_query_datatype::codec::datum_codec

The unified entry for encoding and decoding an evaluable type to / from datum bytes. Datum bytes consists of 1 byte datum flag and variable bytes datum payload.

Traits

ColumnIdDatumEncoder
An encoder to encode a datum storing column id.

DatumFlagAndPayloadEncoder
An encoder to encode a datum, i.e. 1 byte flag + variable bytes payload.

DatumPayloadDecoder
A decoder to decode the payload part of a datum.

DatumPayloadEncoder
An encoder to encode the payload part of a datum.

EvaluableDatumEncoder
An encoder to encode an evaluable type to datum bytes.

RawDatumDecoder

Functions

decode_bytes_datum

decode_date_time_datum

decode_decimal_datum

decode_duration_datum

decode_int_datum

decode_json_datum

decode_real_datum