Trait tidb_query_datatype::codec::datum_codec::DatumPayloadDecoder

pub trait DatumPayloadDecoder: NumberDecoder + CompactByteDecoder + MemComparableByteDecoder + DurationDecoder + TimeDecoder + DecimalDecoder + JsonDecoder + EnumDecoder {
    fn read_datum_payload_i64(&mut self) -> Result<i64> { ... }

    fn read_datum_payload_u64(&mut self) -> Result<u64> { ... }

    fn read_datum_payload_var_i64(&mut self) -> Result<i64> { ... }

    fn read_datum_payload_var_u64(&mut self) -> Result<u64> { ... }

    fn read_datum_payload_f64(&mut self) -> Result<f64> { ... }

    fn read_datum_payload_decimal(&mut self) -> Result<Decimal> { ... }

    fn read_datum_payload_bytes(&mut self) -> Result<Vec<u8>> { ... }

    fn read_datum_payload_compact_bytes(&mut self) -> Result<Vec<u8>> { ... }

    fn read_datum_payload_datetime_int(
        &mut self, 
        ctx: &mut EvalContext, 
        field_type: &FieldType
    ) -> Result<DateTime> { ... }

    fn read_datum_payload_datetime_varint(
        &mut self, 
        ctx: &mut EvalContext, 
        field_type: &FieldType
    ) -> Result<DateTime> { ... }

    fn read_datum_payload_duration_int(
        &mut self, 
        field_type: &FieldType
    ) -> Result<Duration> { ... }

    fn read_datum_payload_duration_varint(
        &mut self, 
        field_type: &FieldType
    ) -> Result<Duration> { ... }

    fn read_datum_payload_json(&mut self) -> Result<Json> { ... }

    fn read_datum_payload_enum_compact_bytes(
        &mut self, 
        field_type: &FieldType
    ) -> Result<Enum> { ... }

    fn read_datum_payload_enum_uint(
        &mut self, 
        field_type: &FieldType
    ) -> Result<Enum> { ... }

    fn read_datum_payload_enum_var_uint(
        &mut self, 
        field_type: &FieldType
    ) -> Result<Enum> { ... }
}

A decoder to decode the payload part of a datum.

The types this decoder outputs are not fully 1:1 mapping to evaluable types.

Provided methods

fn read_datum_payload_i64(&mut self) -> Result<i64>

fn read_datum_payload_u64(&mut self) -> Result<u64>

fn read_datum_payload_var_i64(&mut self) -> Result<i64>

fn read_datum_payload_var_u64(&mut self) -> Result<u64>

fn read_datum_payload_f64(&mut self) -> Result<f64>

fn read_datum_payload_decimal(&mut self) -> Result<Decimal>

fn read_datum_payload_bytes(&mut self) -> Result<Vec<u8>>

fn read_datum_payload_compact_bytes(&mut self) -> Result<Vec<u8>>

fn read_datum_payload_datetime_int(
&mut self,
ctx: &mut EvalContext,
field_type: &FieldType
) -> Result<DateTime>

fn read_datum_payload_datetime_varint(
&mut self,
ctx: &mut EvalContext,
field_type: &FieldType
) -> Result<DateTime>

fn read_datum_payload_duration_int(
&mut self,
field_type: &FieldType
) -> Result<Duration>

fn read_datum_payload_duration_varint(
&mut self,
field_type: &FieldType
) -> Result<Duration>

fn read_datum_payload_json(&mut self) -> Result<Json>

fn read_datum_payload_enum_compact_bytes(
&mut self,
field_type: &FieldType
) -> Result<Enum>

fn read_datum_payload_enum_uint(
&mut self,
field_type: &FieldType
) -> Result<Enum>

fn read_datum_payload_enum_var_uint(
&mut self,
field_type: &FieldType
) -> Result<Enum>

Loading content...