Module tidb_query_datatype::codec::mysql

Re-exports

pub use self::decimal::dec_encoded_len;

pub use self::decimal::Decimal;

pub use self::decimal::DecimalDecoder;

pub use self::decimal::DecimalEncoder;

pub use self::decimal::Res;

pub use self::decimal::RoundMode;

pub use self::duration::Duration;

pub use self::duration::DurationDecoder;

pub use self::duration::DurationEncoder;

pub use self::enums::Enum;

pub use self::enums::EnumRef;

pub use self::json::Json;

pub use self::json::JsonType;

pub use self::set::Set;

pub use self::set::SetRef;

pub use self::time::Time;

pub use self::time::TimeDecoder;

pub use self::time::TimeEncoder;

pub use self::time::TimeType;

Modules

binary_literal

charset

decimal

duration

enums

json
The binary JSON format from MySQL 5.7 is as follows:

set

time

Structs

PathExpression

Enums

ModifyType
ModifyType is for modify a JSON.

Tz
A time zone represented by either offset (i.e. +8) or name (i.e. Asia/Shanghai). In addition, local time zone is also valid.

Constants

DEFAULT_FSP
DEFAULT_FSP is the default digit of fractional seconds part. MySQL use 0 as the default Fsp.

MAX_FSP
MAX_FSP is the maximum digit of fractional seconds part.

MIN_FSP
MIN_FSP is the minimum digit of fractional seconds part.

UNSPECIFIED_FSP
UNSPECIFIED_FSP is the unspecified fractional seconds part.

Traits

JsonDatumPayloadChunkEncoder

JsonDecoder

JsonEncoder

Functions

check_fsp

parse_json_path_expr
Parses a JSON path expression. Returns a PathExpression object which can be used in JSON_EXTRACT, JSON_SET and so on.