Struct tidb_query_datatype::codec::mysql::decimal::Decimal

#[repr(C)]pub struct Decimal {
    int_cnt: u8,
    frac_cnt: u8,
    result_frac_cnt: u8,
    negative: bool,
    word_buf: [u32; 9],
}

Decimal represents a decimal value.

Fields
int_cnt: u8

The number of decimal digits before the point.

frac_cnt: u8

The number of decimal digits after the point.

result_frac_cnt: u8

The number of calculated or printed result fraction digits.

negative: boolword_buf: [u32; 9]

An array of u32 words. A word is an u32 value can hold 9 digits.(0 <= word < wordBase)

Implementations

impl Decimal

pub fn abs(self) -> Res<Decimal>

abs the Decimal into a new Decimal.

pub fn ceil(&self) -> Res<Decimal>

ceil the Decimal into a new Decimal.

pub fn floor(&self) -> Res<Decimal>

floor the Decimal into a new Decimal.

fn new(int_cnt: u8, frac_cnt: u8, negative: bool) -> Decimal

create a new decimal for internal usage.

pub fn is_negative(&self) -> bool

pub fn zero() -> Decimal

Creates a new decimal which is zero.

fn remove_leading_zeroes(&self, prec: u8) -> (usize, u8)

Given a precision count ‘prec’, get:

the index of first non-zero word in self.word_buf to hold the leading ‘prec’ number of digits

the number of remained digits if we remove all leading zeros for the leading ‘prec’ number of digits

fn prepare_buf(&self) -> (Vec<u8>, usize, u8, u8, u8)

Prepare a buf for string output.

pub fn prec_and_frac(&self) -> (u8, u8)

Get the least precision and fraction count to encode this decimal completely.

fn digit_bounds(&self) -> (u8, u8)

digit_bounds returns bounds of decimal digits in the number.

fn do_mini_left_shift(self, shift: u8, beg: u8, end: u8) -> Decimal

do_mini_left_shift does left shift for alignment of data in buffer.

Result fitting in the buffer should be garanted. ‘shift’ have to be from 1 to DIGITS_PER_WORD - 1 (inclusive)

fn do_mini_right_shift(self, shift: u8, beg: u8, end: u8) -> Decimal

do_mini_right_shift does right shift for alignment of data in buffer.

Result fitting in the buffer should be garanted. ‘shift’ have to be from 1 to DIGITS_PER_WORD - 1 (inclusive)

pub fn convert_to(
self,
ctx: &mut EvalContext,
flen: u8,
decimal: u8
) -> Result<Decimal>

convert_to(ProduceDecWithSpecifiedTp in tidb) produces a new decimal according to flen and decimal.

pub fn round(self, frac: i8, round_mode: RoundMode) -> Res<Decimal>

Round rounds the decimal to “frac” digits.

NOTES scale can be negative ! one TRUNCATED error (line XXX below) isn’t treated very logical :(

fn round_with_word_buf_len(
self,
frac: i8,
word_buf_len: u8,
round_mode: RoundMode
) -> Res<Decimal>

fn handle_incr(
res: Res<Decimal>,
int_word_cnt: u8,
frac_words_to: i8,
frac: i8,
frac_word_cnt: u8,
word_buf_len: u8,
round_mode: RoundMode
) -> Res<Decimal>

fn handle_carry(
dec: Res<Decimal>,
to_idx: usize,
frac: i8,
frac_word_to: i8,
int_word_cnt: u8,
word_buf_len: u8
) -> Res<Decimal>

pub fn shift(self, shift: isize) -> Res<Decimal>

shift shifts decimal digits in given number (with rounding if it need), shift > 0 means shift to left shift, shift < 0 means right shift.

In fact it is multiplying on 10^shift.

fn shift_with_word_buf_len(self, shift: isize, word_buf_len: u8) -> Res<Decimal>

pub fn as_i64(&self) -> Res<i64>

as_i64 returns int part of the decimal.

pub fn as_i64_with_ctx(&self, ctx: &mut EvalContext) -> Result<i64>

as_i64_with_ctx returns int part of the decimal.

pub fn as_u64(&self) -> Res<u64>

as_u64 returns int part of the decimal

pub fn from_f64(val: f64) -> Result<Decimal>

pub fn from_bytes(s: &[u8]) -> Result<Res<Decimal>>

Returns a Decimal from a given bytes slice

Notes

An error will be returned if the given input is as follows:

empty string

string which cannot be converted to decimal

fn from_bytes_with_word_buf(s: &[u8], word_buf_len: u8) -> Result<Res<Decimal>>

Returns a Decimal from a given bytes slice buffer and specified buffer length

Notes

An error will be returned if the given input is as follows:

an empty string

a string which cannot be converted to decimal

pub fn approximate_encoded_size(&self) -> usize

Get the approximate needed capacity to encode this decimal.

see also encode_decimal.

fn div(&self, rhs: &Decimal, frac_incr: u8) -> Option<Res<Decimal>>

pub fn is_zero(&self) -> bool