Struct tidb_query_datatype::codec::mysql::duration::Duration

#[repr(C)]pub struct Duration {
    nanos: i64,
    fsp: u8,
}

Fields
nanos: i64fsp: u8
Implementations

impl Duration

pub fn is_neg(self) -> bool

pub fn hours(self) -> u32

pub fn minutes(self) -> u32

pub fn secs(self) -> u32

pub fn subsec_micros(self) -> u32

Returns the fractional part of Duration in microseconds.

pub fn subsec_nanos(self) -> u32

Returns the fractional part of Duration in nanoseconds.

pub fn fsp(self) -> u8

pub fn maximize_fsp(self) -> Self

pub fn to_secs(self) -> i64

pub fn to_secs_f64(self) -> f64

Returns the number of seconds contained by this Duration as f64. The returned value does include the fractional (nanosecond) part of the duration.

pub fn to_millis(self) -> i64

pub fn to_micros(self) -> i64

pub fn to_nanos(self) -> i64

pub fn zero() -> Duration

Returns the identity element of Duration

pub fn is_zero(self) -> bool

Returns true if self is equal to the additive identity.

pub fn abs(self) -> Self

Returns the absolute value of Duration

pub fn from_secs(secs: i64, fsp: i8) -> Result<Duration>

pub fn from_millis(millis: i64, fsp: i8) -> Result<Duration>

pub fn from_micros(micros: i64, fsp: i8) -> Result<Duration>

pub fn from_nanos(nanos: i64, fsp: i8) -> Result<Duration>

pub fn new_from_parts(
neg: bool,
hour: u32,
minute: u32,
second: u32,
nanos: u32,
fsp: i8
) -> Result<Duration>

pub fn parse(ctx: &mut EvalContext, input: &[u8], fsp: i8) -> Result<Duration>

Parses the time from a formatted string with a fractional seconds part, returns the duration type Time value. See: http://dev.mysql.com/doc/refman/5.7/en/fractional-seconds.html

pub fn round_frac(self, fsp: i8) -> Result<Self>

Rounds fractional seconds precision with new FSP and returns a new one. We will use the “round half up” rule, e.g, >= 0.5 -> 1, < 0.5 -> 0, so 10:10:10.999999 round with fsp: 1 -> 10:10:11.0 and 10:10:10.000000 round with fsp: 0 -> 10:10:11

pub fn checked_add(self, rhs: Duration) -> Option<Duration>

Checked duration addition. Computes self + rhs, returning None if overflow occurred.

pub fn checked_sub(self, rhs: Duration) -> Option<Duration>

Checked duration subtraction. Computes self - rhs, returning None if overflow occurred.

fn format(self, sep: &str) -> String

pub fn to_numeric_string(self) -> String

Converts a Duration to printable numeric string representation

pub fn from_i64(ctx: &mut EvalContext, n: i64, fsp: i8) -> Result<Duration>