Struct tidb_query_datatype::codec::mysql::json::Json

pub struct Json {
    type_code: JsonType,
    pub value: Vec<u8>,
}

Json implements type json used in tikv by Binary Json. The Binary Json format from MySQL 5.7 is in the following link: (https://github.com/mysql/mysql-server/blob/5.7/sql/json_binary.h#L52) The only difference is that we use large object or large array for the small corresponding ones. That means in our implementation there is no difference between small object and big object, so does array.

Fields
type_code: JsonTypevalue: Vec<u8>

The binary encoded json data in bytes

Implementations

impl Json

pub fn merge<'a>(bjs: Vec<JsonRef<'a>>) -> Result<Json>

merge is the implementation for JSON_MERGE in mysql https://dev.mysql.com/doc/refman/5.7/en/json-modification-functions.html#function_json-merge

The merge rules are listed as following:

adjacent arrays are merged to a single array;

adjacent object are merged to a single object;

a scalar value is autowrapped as an array before merge;

an adjacent array and object are merged by autowrapping the object as an array.

See MergeBinary() in TiDB json/binary_function.go

impl Json

pub fn new(tp: JsonType, value: Vec<u8>) -> Self

Creates a new JSON from the type and encoded bytes

pub fn get_type(&self) -> JsonType

Returns the JSON type

pub fn from_string(s: String) -> Result<Self>

Creates a string JSON from a String

pub fn from_str_val(s: &str) -> Result<Self>

Creates a string JSON from a &str

pub fn from_bool(b: bool) -> Result<Self>

Creates a literal JSON from a bool

pub fn from_u64(v: u64) -> Result<Self>

Creates a number JSON from a u64

pub fn from_f64(v: f64) -> Result<Self>

Creates a number JSON from a f64

pub fn from_i64(v: i64) -> Result<Self>

Creates a number JSON from an i64

pub fn from_ref_array(array: Vec<JsonRef>) -> Result<Self>

Creates a array JSON from a collection of JsonRef

pub fn from_array(array: Vec<Json>) -> Result<Self>

Creates a array JSON from a collection of Json

pub fn from_kv_pairs<'a>(entries: Vec<(&[u8], JsonRef<'a>)>) -> Result<Self>

Creates a object JSON from key-value pairs

pub fn from_object(map: BTreeMap<String, Json>) -> Result<Self>

Creates a object JSON from key-value pairs in BTreeMap

pub fn none() -> Result<Self>

Creates a null JSON

pub fn as_ref(&self) -> JsonRef

Returns a JsonRef points to the starting of encoded bytes