Struct tidb_query_datatype::codec::mysql::json::JsonRef

pub struct JsonRef<'a> {
    type_code: JsonType,
    value: &'a [u8],
}

Represents a reference of JSON value aiming to reduce memory copy.

Fields
type_code: JsonTypevalue: &'a [u8]
Implementations

impl<'a> JsonRef<'a>

pub fn array_get_elem(&self, idx: usize) -> Result<JsonRef<'a>>

Gets the ith element in JsonRef

See arrayGetElem() in TiDB json/binary.go

pub fn object_get_key(&self, i: usize) -> &'a [u8]

Return the ith key in current Object json

See arrayGetElem() in TiDB json/binary.go

pub fn object_get_val(&self, i: usize) -> Result<JsonRef<'a>>

Returns the JsonRef of ith value in current Object json

See arrayGetElem() in TiDB json/binary.go

pub fn object_search_key(&self, key: &[u8]) -> Option<usize>

Searches the value index by the give key in Object.

See objectSearchKey() in TiDB json/binary_function.go

pub fn val_entry_get(&self, val_entry_off: usize) -> Result<JsonRef<'a>>

Gets the value (JsonRef) by the given offset of the value entry

See arrayGetElem() in TiDB json/binary.go

pub(super) fn as_ptr(&self) -> *const u8

Returns a raw pointer to the underlying values buffer.

pub(super) fn as_literal(&self) -> Result<u8>

Returns the literal value of JSON document

pub fn binary_len(&self) -> usize

Returns the encoding binary length of self

impl<'a> JsonRef<'a>

fn get_precedence(&self) -> i32

fn as_f64(&self) -> Result<f64>

impl<'a> JsonRef<'a>

fn encoded_len(&self) -> usize

impl<'a> JsonRef<'a>

pub fn depth(&self) -> Result<i64>

Returns maximum depth of JSON document

impl<'a> JsonRef<'a>

pub fn extract(&self, path_expr_list: &[PathExpression]) -> Result<Option<Json>>

extract receives several path expressions as arguments, matches them in j, and returns the target JSON matched any path expressions, which may be autowrapped as an array. If there is no any expression matched, it returns None.

See Extract() in TiDB json.binary_function.go

impl<'a> JsonRef<'a>

pub fn keys(&self, path_expr_list: &[PathExpression]) -> Result<Option<Json>>

Evaluates a (possibly empty) list of values and returns a JSON array containing those values specified by path_expr_list

impl<'a> JsonRef<'a>

fn len(&self) -> i64

pub fn json_length(
&self,
path_expr_list: &[PathExpression]
) -> Result<Option<i64>>

json_length is the implementation for JSON_LENGTH in mysql https://dev.mysql.com/doc/refman/5.7/en/json-attribute-functions.html#function_json-length

impl<'a> JsonRef<'a>

pub fn modify(
&self,
path_expr_list: &[PathExpression],
values: Vec<Json>,
mt: ModifyType
) -> Result<Json>

Modifies a Json object by insert, replace or set. All path expressions cannot contain * or ** wildcard. If any error occurs, the input won't be changed.

See Modify() in TiDB json/binary_function.go

impl<'a> JsonRef<'a>

pub fn remove(&self, path_expr_list: &[PathExpression]) -> Result<Json>

Removes elements from Json, All path expressions cannot contain * or ** wildcard. If any error occurs, the input won't be changed.

impl<'a> JsonRef<'a>

pub fn json_type(&self) -> &'static [u8]

json_type is the implementation for https://dev.mysql.com/doc/refman/5.7/en/json-attribute-functions.html#function_json-type

impl<'a> JsonRef<'a>

pub fn unquote(&self) -> Result<String>

unquote recognizes the escape sequences shown in: https://dev.mysql.com/doc/refman/5.7/en/json-modification-functions.html# json-unquote-character-escape-sequences

See Unquote() in TiDB json/binary_function.go

impl<'a> JsonRef<'a>

pub fn new(type_code: JsonType, value: &[u8]) -> JsonRef<'_>

pub fn to_owned(&self) -> Json

Returns an owned Json via copying

pub fn get_type(&self) -> JsonType

Returns the JSON type

pub fn value(&self) -> &'a [u8]

Returns the underlying value slice

pub(crate) fn get_u64(&self) -> u64

pub(crate) fn get_i64(&self) -> i64

pub(crate) fn get_double(&self) -> f64

pub(crate) fn get_elem_count(&self) -> usize

pub(crate) fn get_literal(&self) -> Option<bool>

pub(crate) fn get_str_bytes(&self) -> Result<&'a [u8]>

pub(crate) fn get_str(&self) -> Result<&'a str>

pub(crate) fn is_zero(&self) -> bool