Module tidb_query_datatype::codec::mysql::time::parser

Functions

adjust_year
If a two-digit year encountered, add an offset to it. 99 -> 1999 20 -> 2020

bytes_to_u32

digit1
Match at least one digit and return the rest of the slice.

parse

parse_frac
Try to parse a fractional part from input with fsp, round the result if round is true. NOTE: This function assumes that fsp is in range: [0, 6].

parse_from_decimal

parse_from_i64
Try to parse a i64 into a Time with the given type and fsp

parse_whole
Try to parse a datetime string input without fractional part and separators. return an array that stores [year, month, day, hour, minute, second, 0]

space1
Match at least one space and return the rest of the slice.

split_components
We assume that the input is trimmed and is not empty.