Module tidb_query_datatype::codec::row::v2::encoder_for_test

This encoder module is only used for test, so the implementation is very straightforward.

According to https://github.com/pingcap/tidb/blob/master/docs/design/2018-07-19-row-format.md

The row format is:

version	flag	number_of_non_null_columns	number_of_null_columns	non_null_column_ids	null_column_ids	value_offsets	values

length about each field:

version: 1 byte

flag: 1 byte, when there's id greater than 255 or the total size of the values is greater than 65535 , value is 1, otherwise 0

number of non-null values: 2 bytes

number of null values: 2 bytes

non-null column ids: when flag == 1 (big), id is 4 bytes, otherwise 1 byte

null column ids: when flag == 1 (big), id is 4 bytes, otherwise 1 byte

non-null values offset: when big, offset is 4 bytes, otherwise 2 bytes

Structs

Column

Constants

MAX_I8

MAX_I16

MAX_I32

MAX_U8

MAX_U16

MAX_U32

MIN_I8

MIN_I16

MIN_I32

Traits

RowEncoder

ScalarValueEncoder