Struct tidb_query_datatype::codec::chunk::Column

pub struct Column {
    length: usize,
    null_cnt: usize,
    null_bitmap: Vec<u8>,
    var_offsets: Vec<usize>,
    data: Vec<u8>,
    fixed_len: usize,
}

Column stores the same column data of multi rows in one chunk.

Fields
length: usizenull_cnt: usizenull_bitmap: Vec<u8>var_offsets: Vec<usize>data: Vec<u8>fixed_len: usize
Implementations

impl Column

pub fn new(field_type_tp: FieldTypeTp, init_cap: usize) -> Column

Create the column with a specified type and capacity.

pub fn from_raw_datums(
field_type: &FieldType,
raw_datums: &BufferVec,
logical_rows: &[usize],
ctx: &mut EvalContext
) -> Result<Self>

pub fn from_vector_value(
field_type: &FieldType,
v: &VectorValue,
logical_rows: &[usize]
) -> Result<Self>

pub fn get_datum(
&self,
idx: usize,
field_type: &dyn FieldTypeAccessor
) -> Result<Datum>

Get the datum of one row with the specified type.

pub fn append_datum(&mut self, data: &Datum) -> Result<()>

Append datum to the column.

pub fn new_fixed_len(element_len: usize, init_cap: usize) -> Column

Create a column with a fixed element length.

pub fn new_var_len_column(init_cap: usize) -> Column

Create a column with variant element length.

fn is_fixed(&self) -> bool

Return whether the column has a fixed length or not.

pub fn get_fixed_len(&self) -> usize

Return the column's fixed length.

pub fn reset(&mut self)

Reset the column

pub fn is_null(&self, row_idx: usize) -> bool

Return whether the datum for the row is null or not.

fn append_null_bitmap(&mut self, on: bool)

Update the null bitmap and count when append a datum. on is false means the datum is null.

pub fn append_null(&mut self)

Append null to the column.

fn finish_append_fixed(&mut self)

Called when the fixed datum has been appended.

fn finished_append_var(&mut self)

Called when the variant datum has been appended.

pub fn append_i64(&mut self, v: i64) -> Result<()>

Append i64 datum to the column.

pub fn append_i64_datum(&mut self, src_datum: &[u8]) -> Result<()>

pub fn get_i64(&self, idx: usize) -> Result<i64>

Get the i64 datum of the row in the column.

pub fn append_u64(&mut self, v: u64) -> Result<()>

Append u64 datum to the column.

pub fn append_u64_datum(&mut self, src_datum: &[u8]) -> Result<()>

pub fn get_u64(&self, idx: usize) -> Result<u64>

Get the u64 datum of the row in the column.

pub fn append_f64(&mut self, v: f64) -> Result<()>

Append a f64 datum to the column.

pub fn append_f64_datum(&mut self, src_datum: &[u8]) -> Result<()>

pub fn append_f32(&mut self, v: f32) -> Result<()>

Append a f32 datum to the column.

pub fn append_f32_datum(&mut self, src_datum: &[u8]) -> Result<()>

pub fn get_f64(&self, idx: usize) -> Result<f64>

Get the f64 datum of the row in the column.

pub fn get_f32(&self, idx: usize) -> Result<f32>

Get the f32 datum of the row in the column.

pub fn append_bytes(&mut self, byte: &[u8]) -> Result<()>

Append a bytes datum to the column.

pub fn append_bytes_datum(&mut self, src_datum: &[u8]) -> Result<()>

pub fn get_bytes(&self, idx: usize) -> &[u8]

Get the bytes datum of the row in the column.

pub fn append_time(&mut self, t: Time) -> Result<()>

Append a time datum to the column.

pub fn append_time_datum(
&mut self,
src_datum: &[u8],
ctx: &mut EvalContext,
field_type: &FieldType
) -> Result<()>

pub fn get_time(&self, idx: usize) -> Result<Time>

Get the time datum of the row in the column.

pub fn append_duration(&mut self, d: Duration) -> Result<()>

Append a duration datum to the column.

pub fn append_duration_datum(&mut self, src_datum: &[u8]) -> Result<()>

pub fn get_duration(&self, idx: usize, fsp: isize) -> Result<Duration>

Get the duration datum of the row in the column.

pub fn append_decimal(&mut self, d: &Decimal) -> Result<()>

Append a decimal datum to the column.

pub fn append_decimal_datum(&mut self, src_datum: &[u8]) -> Result<()>

pub fn get_decimal(&self, idx: usize) -> Result<Decimal>

Get the decimal datum of the row in the column.

pub fn append_json(&mut self, j: JsonRef<'_>) -> Result<()>

Append a json datum to the column.

pub fn append_json_datum(&mut self, src_datum: &[u8]) -> Result<()>

Append a json datum in raw bytes to the column.

pub fn get_json(&self, idx: usize) -> Result<Json>

Get the json datum of the row in the column.

pub fn len(&self) -> usize

Return the total rows in the column.

pub fn is_empty(&self) -> bool